Before major version increment in API
This is another silly housekeeping update, triggered by a discovery
that there will need to be a number of coordinated changes soon.

Concretely, the API::dynamic command (used in 1 place in total!)
would be more useful if requirement lists were replaced by
requirement scripts. This is no big deal, since code can be written
to support both formalisations, but is technically an incompatible
change and requires a major version number update, from 1 to 2. While
I'm at that, I might as well do some more cleanup in the package.
Which I will start *after* this commit.