Another random collection: Planarity, roadmap, Sqlite fix, documentation tweaks.
This commit gathers changes from 2022--2023, so there is not a common
theme to them.

One group of changes early in the period is mostly documentation
improvements, especially in cmplutil.tex. This work also suggested some
new options for network layout and rules export, in network2.dtx or
cmplutil2.dtx as appropriate.

Minor: Having been reminded about the concept of integrals in Hopf
algebras, and this time actually getting what they are about, I may have
started a problem setup for "Hopf algebra with integral" (docstrip module
<integral>).

It looks like there has been some work on network::rich::construct. I'm
not sure if it got finished.

One feature whose implementation has long been overdue is planarity
checking, now provided via network::pure::genus or
network::pure::open_boundaries (depending on whether order of external
edges is taken as fixed or not), and used in
network::wfb::ambiguities_by_planarity. Nothing actually uses that yet,
but it could be provided as a variant in the completion utility version 1
demo. (Imposing a planarity constraint makes no mathematical sense for
that axiom system, but it should produce a noticable effect, which is what
we are after for a demo.)

Poking around in the source code has lead to improved explanations here
and there.

Notable: For those afflicted by Apple's sqlite3 package segfaulting,
cmplutil2.dtx now offers a new version of the fix for this, in docstrip
module <AppleSqliteFix>. LIVE TESTING NEEDED, which means editing the
.ins file to include this.

Finally, there is now a roadmap (roadmap.md) to future development options.